ORDER-0: buisness logic fixes

+++ src/main/java/com/innowise/orderservice/service/impl/ItemServiceImpl.java
@@ -1,11 +1,13 @@
 package com.innowise.orderservice.service.impl;
 
+import com.innowise.orderservice.exception.ItemInUseException;
 import com.innowise.orderservice.exception.ItemNotFoundException;
 import com.innowise.orderservice.mapper.ItemMapper;
 import com.innowise.orderservice.model.dto.item.ItemRequestDto;
 import com.innowise.orderservice.model.dto.item.ItemResponseDto;
 import com.innowise.orderservice.model.entity.Item;
 import com.innowise.orderservice.repository.ItemRepository;
+import com.innowise.orderservice.repository.OrderItemRepository;
 import com.innowise.orderservice.service.ItemService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -18,6 +20,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class ItemServiceImpl implements ItemService {
     private final ItemRepository itemRepository;
     private final ItemMapper itemMapper;
+    private final OrderItemRepository orderItemRepository;
 
     @Override
     @Transactional
@@ -51,6 +54,9 @@ public class ItemServiceImpl implements ItemService {
     @Override
     @Transactional
     public void deleteItemById(Long id) {
+        if (orderItemRepository.existsByItemId(id)) {
+            throw new ItemInUseException("Cannot delete an item that is in order");
+        }
         Item item = getExistingItem(id);
         itemRepository.delete(item);
     }
